Change DB Lat raise range
If I use the same size range as for curls I can keep using the same
dumbbell weights on both every day, which saves a lot of time!

I'm sure this will break down eventually, but I see no reason not to
ride the convenience as long as I can.