Up dead hang time
I did a test, and can manage a 60s hang, but it was a little intense.
I'm going to do 45s. I'm dropping the volume - I think I was honestly
overdoing it a little. Deepseek seems to think I'll progress fine with 2
sets, so lets do that!

If I don't feel good progress I can up the volume later, and I'm
unlikely to regret the extra recovery this buys me even if I take an
extra week getting to 60s hangs.