Drop Isometric 90-90 press
This just doesn't feel like it's doing much, and certainly not hitting
any weakness I can think of.

If I remember right, I added it as part of a push towards more dynamic
stretches aiming at getting that loose and flexible feeling for my hips
that I've been enjoying in my shoulders. The 90-90 switches and the
cossack squats are doing great. I think I just went a little overboard,
and added a stretch that I don't really understand and doesn't do much.
Let's not do that!