Tweak monday exercise order
I had the dips near the bottom on the premise that I'd be doing the
other heavy lifts in the morning, and do the dips in a separate session.
But in practice, I've just been doing them early, so lets put them where
I actually do them (before the accessory work).