Add optional exercises and drop Bretzel
The Bretzel wasn't doing much, and the pigeon pose is a good substitute.

Sun Salutations provide a nice full body movement warm up, and probably
add a lot of value here, but it's also a substantial non-essential
adition to the routine.

Squat holds are similarly non-vital, the core value being covered by the
cossack squats and split squats.

Having some optional exercises makes it possible to keep a solid tight
core workout, but include some of the would-be-nices without
compromising efficiency.