Add "Glute Bias" note to BSS
I'd like to work towards a glute bias on these. Longer stance, plus
leaning forward seems to be the trick.

That said, I'm going to be mostly focused on just doing solid split
squats for the next week or two without worrying about tweaking the form
for a glute bias. I'm putting this note in place so that I start
thinking about it for future form tweaks.