Re-arrange mobility work for my poor glutes
Doing spiderman lunges and cossack squats the day before heavy bulgarian
split squats was apparently a mistake. My glutes are so unhappy.

I can consider re-arranging again in a few weeks once the exercise is
less novel, but for now, let's put the cossacks and spiderman lunges
*after* the heavy lifts. Maybe I'll like the new arrangement and keep
it.

I haven't changed the exercises at all, just the days and order.