Set one-armed pushup volume to maintenance level
I'll be doing 1x5 for the foreseeable future since I don't really want
to progress this, and it isn't an integral part of my progression plan,
but I like the functional mobility it gives me so I don't want to drop
it altogether.

Eventually, I may consider some sort of progressions towards fancier
bodyweight exercises. I dream of one day doing planche pushups. But I
think that for now, I should continue forcusing on the bench press for
pushing strength, and maybe return to the bodyweight side when my
bodyweight is a little reduced.