Tweak BSS routine
Instead of heavy clusters, I need some volume and skillwork to build up
the small muscles.

My glute medius in particular probably needs some attention, so for the
next could weeks I'll do single leg glute bridge holds in place of the
cossack squats to help move that along. The plan is to go back to
cossack squats after a few weeks, and also either drop the Friday
bodyweight BSS or replace it with hip thursts if I'm feeling like
continuing to push the glute focus.