Rework ankle mobility exercises
My mobility improves in the short term when I do these stretches, but
I'm seeing no measurable gains compared to a month ago to my
dorsiflexion range.

My current best theory is that my calves are dominant over my tibialis
anterior which is preventing proper gliding and rotation of the tibia.
I'm not sure this is right, but I've tried more or less everything else.
The next step if this fails is either to give up on ankle mobility
improving or talk to a PT.