Drop wednesday band pull-aparts
These were useful when this was a new exercise to get my muscles up to
speed. At this point, I'm seeing diminishing returns on this guy, and
doing the 3-way band pull-aparts once a week should be more than enough
for prehab, which was my main reason for adding this exercise.