Temporarily reduce work for tennis elbow
I don't actually have tennis elbow yet. Just some mild discomfort for a
bit post heavy lifting. Let's keep it that way! I'll reevaluate in 2
weeks, and reintroduce the missing exercises and volume if the
discomfort is gone!

This should also give me a couple weeks for my wrist extensions to do
some work, and help prevent this happening in the future. The hope here
is that if I strengthen my forearms and at the same time let the tendons
recover by the time I reintroduce the volume, I'll be both stronger and
recovered, and the tendons will have had a chance to adapt also.