Update ankle work, re-organize mobility
I've dropped the Tibialis raises and instead will be doing banded ankle
distractions and weighted ankle mobs. That should be the only
substantive change here, the rest is just re-ordering for convenience,
and swapping of some days for balance. Now the last bunch of exercises
in the list is my temporary prehab stuff.