Reduce pancake frequency and rearrange
I think that at this point if I want to progress more on the pancake I
probably need to give it more time than I'm currently willing. 2x/week
should be good for maintenance, and I'll try that for now. I may drop it
and I may eventually decide to do a more concentrated push, but this is
good for now!